Behavior matching
The strip_tags function on PHP removes chars after a broken tag (i.e. "<b" without the closing char ">") til the end of the string and removes the "<" char when its the last on the string. 
The proposed change removes the redundance on range matching regex string and adds a ? allowing to get that broken tags. Besides, it also checks the last char on the string and removes it when necessary.